+++ internal/servers/controller/worker_tls_config.go
@@ -7,10 +7,12 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"net"
 	"strings"
 
 	"github.com/hashicorp/boundary/internal/cmd/base"
+	"github.com/hashicorp/boundary/internal/servers"
 	wrapping "github.com/hashicorp/go-kms-wrapping"
 	"google.golang.org/protobuf/proto"
 )
@@ -20,16 +22,30 @@ type workerAuthEntry struct {
 	conn net.Conn
 }
 
+// validateWorkerTls is called by the Go TLS stack with client info. It calls
+// v1WorkerAuthConfig to validate the encryption and ensure the nonce has not
+// been replayed, then stores connection information into the auth cache, which
+// is used by the intercepting listener to log information.
 func (c Controller) validateWorkerTls(hello *tls.ClientHelloInfo) (*tls.Config, error) {
 	for _, p := range hello.SupportedProtos {
 		switch {
 		case strings.HasPrefix(p, "v1workerauth-"):
 			tlsConf, workerInfo, err := c.v1WorkerAuthConfig(hello.SupportedProtos)
 			if err == nil {
-				// Set the info we need to prevent replays
-				c.workerAuthCache.Set(workerInfo.ConnectionNonce, &workerAuthEntry{
+				// Store info that can be retrieved in the intercepting
+				// listener.
+				//
+				// NOTE: This is not a load-or-store because we need to ensure
+				// replays can't happen across controllers, hence storing in the
+				// database in v1WorkerAuthConfig. Although this could serve as
+				// a secondary check, a load-or-store here keeps the test from
+				// ensuring that it is rejected at the database level via the
+				// nonce unique constraint, which is by far the more important
+				// thing to validate (although there are DB-level tests for that
+				// too).
+				c.workerAuthCache.Store(workerInfo.ConnectionNonce, &workerAuthEntry{
 					WorkerAuthInfo: workerInfo,
-				}, 0)
+				})
 			}
 			return tlsConf, err
 		}
@@ -37,6 +53,13 @@ func (c Controller) validateWorkerTls(hello *tls.ClientHelloInfo) (*tls.Config, 
 	return nil, nil
 }
 
+// v1WorkerAuthConfig:
+//
+// * Reads the information from ALPN, combining across protos
+// * Combines them to form the encrypted auth information
+// * Validates it by decrypting against the configured KMS
+// * Ensures that the nonce is unique to prevent replay attacks
+// * Returns the shared TLS configuration that is used to establish the connection
 func (c Controller) v1WorkerAuthConfig(protos []string) (*tls.Config, *base.WorkerAuthInfo, error) {
 	var firstMatchProto string
 	var encString string
@@ -67,6 +90,15 @@ func (c Controller) v1WorkerAuthConfig(protos []string) (*tls.Config, *base.Work
 	info := new(base.WorkerAuthInfo)
 	if err := json.Unmarshal(marshaledInfo, info); err != nil {
 		return nil, nil, err
+	}
+
+	// Check for replays
+	serversRepo, err := c.ServersRepoFn()
+	if err != nil {
+		return nil, nil, fmt.Errorf("unable to fetch servers repo: %w", err)
+	}
+	if err := serversRepo.AddNonce(c.baseContext, info.ConnectionNonce, servers.NoncePurposeWorkerAuth); err != nil {
+		return nil, nil, fmt.Errorf("unable to add connection nonce to database: %w", err)
 	}
 
 	rootCAs := x509.NewCertPool()